Add validation for numeric parameter

+++ app/controllers/api/products_controller.rb
@@ -7,14 +7,26 @@ module Api
       end
 
       def show
+        if !is_number?(params[:id])
+          render json: {status:'ERROR', message:'Product id is not a numeric value', data:params[:id]}, status: :unprocessable_entity
+          return;
+        end
         product = Product.find(params[:id]);
         render json: {status:'SUCCESS', message:'Product loaded', data:product}, status: :ok
       end
 
       def destroy
+        if !is_number?(params[:id])
+          render json: {status:'ERROR', message:'Product id is not a numeric value', data:params[:id]}, status: :unprocessable_entity
+          return;
+        end
         product = Product.find(params[:id]);
         product.destroy
         render json: {status:'SUCCESS', message:'Product deleted', data:product}, status: :ok
+      end
+
+      def is_number?(obj)
+        (obj.to_s == obj.to_i.to_s) || (obj.to_s == obj.to_f.to_s)
       end
     end
   end
